fix bug "An incomplete coupling created when no CAD documents were opened before executing the macro" in coupling.FCMacro
Some parts of the macro refered to the global variable "document", which was
set to None if no CAD documents were opened. The global variable "document"
remained None until the macro was closed and executed again.

After the bug fix, the global document is removed completely. The document
information is updated each time the user clicks "OK" in
"Create coupling" dialog.